training: don't round norm weights to compute dtype in the liger path (parity-gate FAIL root cause) (#126)
First box run of the liger grad-parity gate FAILed with exactly the
signature of systematic drift, not corruption: loss rel 6.4e-3, worst
adapter at layer 1 (cos 0.9818, rel 0.20). Root cause: the _norm liger
branch cast the base's fp16 norm weight to the bf16 compute dtype (7 vs 10
mantissa bits, up to ~0.4% per element) before the kernel call, while the
torch reference path upcasts the original fp16 values via w.float().
casting_mode="gemma" upcasts W to fp32 inside the kernel and liger has no
X/W dtype-match assert, so the cast was never needed -- move the weight to
x's device but keep its own dtype.

Also make the gate's report diagnostic: print how many adapters fall
outside tolerance, the median cosine/rel, and the 5 worst offenders, so a
handful of deep-layer outliers (numerics gap) reads differently from a
blown-out list (corrupted backward).

+++ exllamav3/training/native_llama.py
@@ -480,10 +480,16 @@ class NativeLlamaQLoRA(nn.Module):
             if ops is not None:
                 w = spec["weight"]
                 if w is not None:
-                    # Move the norm weight to x's device too (not just dtype): under
-                    # --parallel split a block's weights and activations share a card,
-                    # but the Triton kernel below launches on the *current* CUDA device.
-                    w = w.to(device=x.device, dtype=x.dtype)
+                    # Move the norm weight to x's device (under --parallel split a
+                    # block's weights and activations share a card, but the Triton
+                    # kernel launches on the *current* CUDA device) -- but keep the
+                    # weight's OWN dtype. casting_mode="gemma" upcasts W to fp32
+                    # inside the kernel, so passing the original (fp16) weight
+                    # reproduces the torch path's w.float() exactly; casting to
+                    # x.dtype first (bf16, 7 mantissa bits vs fp16's 10) rounded
+                    # every norm weight by up to ~0.4% on the liger side only --
+                    # the systematic drift the parity gate measured (Session 12).
+                    w = w.to(device=x.device)
                     # The frozen base weight is an inference tensor (the EXL3 model
                     # loads under @torch.inference_mode), and Liger's autograd Function
                     # saves W for backward -- which forbids inference tensors. When the
